+++ server/api/src/main/java/vook/server/api/devhelper/app/InitService.java
@@ -6,8 +6,8 @@ import org.springframework.transaction.annotation.Transactional;
 import vook.server.api.domain.demo.model.DemoTerm;
 import vook.server.api.domain.demo.model.DemoTermRepository;
 import vook.server.api.domain.demo.model.DemoTermSynonymRepository;
+import vook.server.api.domain.template_vocabulary.logic.TemplateVocabularyCreateCommand;
 import vook.server.api.domain.template_vocabulary.logic.TemplateVocabularyLogic;
-import vook.server.api.domain.template_vocabulary.logic.dto.TemplateVocabularyCreateCommand;
 import vook.server.api.domain.template_vocabulary.model.TemplateTermRepository;
 import vook.server.api.domain.template_vocabulary.model.TemplateVocabularyRepository;
 import vook.server.api.domain.template_vocabulary.model.TemplateVocabularyType;
